Fully working code on linux
Has the finished multi grid on it (awsome radar as well :P ). To make work on mac, need to change *ALL* the relative paths to absolute ones (including gnuplot). Includes measuring the time to run as well as the processor ticks and number of loops. Needs to have more graphs inputted... Interact only works when the code is run through qtcreator for some reason.